refactor(music): extract pure music utilities and inject NowPlayingManager
MusicPlayerHelper was a Kotlin object that mixed pure utilities
(formatTime, isUrl, adjustTrackPlayingTimes, deriveDeleteDelayFromTrack)
with stateful playback logic and an embedded NowPlayingManager. Two
problems followed: VoiceEventHandler reached into
MusicPlayerHelper.nowPlayingManager directly as a hidden dependency,
and the pure helpers were untestable in isolation from the helper's
playback machinery.

Move the four pure functions to bot.toby.util.MusicFormat and update
each call site to import them from there. Expose the helper's existing
NowPlayingManager instance as a Spring @Bean (MusicBeans) so handlers
can inject the same instance the helper uses. VoiceEventHandler now
takes NowPlayingManager directly via its constructor.

Stateful methods (playUserIntro, skipTracks, stopSong,
changePauseStatusOnTrack, nowPlaying, resetMessages) stay on
MusicPlayerHelper for now; promoting them to a managed service would
ripple through TrackScheduler and every command/button that imports
them statically, and is best left to a follow-up.
